Rename question-related files
