+++ src/server/models/index.ts
@@ -1,4 +1,4 @@
-import * as questions from "./questions";
+import * as questions from "./question";
 import * as answers from "./answers";
 import * as users from "./user";
